add Softplan company
They are one of the largest software developers in the country and serve several segments with an extreme impact on society, with solutions that are a benchmark for digital transformation in the public and private sectors. Currently, they are the biggest company in Brazil in the Construction Industry field.
They use a remote-first approach, but hybrid work is also an option depending on what you're working with.